test-requirements.txt
testrepository
testtools
python-heatclient
python-muranoclient
python-keystoneclient
requests
